Rename logItem to logMessageItem
