+++ src/components/logdisplay/LogMessageItem.vue
